+++ components/lookup/index.jsx
@@ -52,6 +52,10 @@ const defaultFilter = (term, item) => {
 
 /**
  * Lookup is an advanced inline search form. The lookup can parse through single or multi scoped datasets. The parsed dataset can be filtered by single or multi option selects.
+ *
+ * This component is wrapped in a [higher order component to listen for clicks outside itself](https://github.com/kentor/react-click-outside) and thus requires use of `ReactDOM`.
+ *
+ * This component may use a portalMount (a disconnected React subtree mount) within an absolutely positioned DOM node created with [Drop](http://github.hubspot.com/drop/).
  */
 const Lookup = React.createClass({
 	displayName: LOOKUP,