+++ examples/date-picker/custom-input.jsx
@@ -1,13 +1,23 @@
 /* eslint-disable no-console, react/prop-types */
 import React from 'react';
 import Datepicker from '~/components/date-picker';
+import Input from '~/components/forms/input';
 
 const Example = React.createClass({
 	displayName: 'DatepickerExample',
 
+	getInitialState () {
+		return {
+			isOpen: false
+		};
+	},
+
 	render () {
 		return (
 			<Datepicker
+				isOpen={this.state.isOpen}
+				onRequestClose={() => { this.setState({ isOpen: false }); }}
+				onRequestOpen={() => { this.setState({ isOpen: true }); }}
 				onChange={(event, data) => {
 					if (this.props.action) {
 						const dataAsArray = Object.keys(data).map((key) => data[key]);
@@ -16,15 +26,9 @@ const Example = React.createClass({
 						console.log('onChange', event, data);
 					}
 				}}
-				onCalendarFocus={(event, data) => {
-					if (this.props.action) {
-						const dataAsArray = Object.keys(data).map((key) => data[key]);
-						this.props.action('onCalendarFocus')(event, data, ...dataAsArray);
-					} else if (console) {
-						console.log('onCalendarFocus', event, data);
-					}
-				}}
-			/>
+			>
+				<Input placeholder="With custom Input" value="" />
+			</Datepicker>
 		);
 	}
 });